chore: move tests, condense README
